Clean codes

+++ src/main/java/com/academy/billing/service/BillingServiceImpl.java
@@ -1,5 +1,6 @@
 package com.academy.billing.service;
 
+import com.academy.billing.exception.RecordNotFoundException;
 import com.academy.billing.model.Billing;
 import com.academy.billing.repository.BillingRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,54 +13,51 @@ import java.util.Optional;
 @Service
 public class BillingServiceImpl implements BillingService {
 
-
-
     @Autowired
     BillingRepository billingRepository;
 
 
-
     @Override
-    public Page<Billing> findAllAccounts(Pageable pageable) {
-        Page<Billing> pageBilling = billingRepository.findAll(pageable);
-        return pageBilling;
+    public Billing saveBilling(Billing billing) {
+        return billingRepository.save(billing);
     }
 
-
+    @Override
+    public Page<Billing> findAllBillings(Pageable pageable) {
+        return billingRepository.findAll(pageable);
+    }
 
     @Override
-    public Billing findAccountById(Long id) {
+    public Billing findBillingById(Long id) throws RecordNotFoundException {
         Optional<Billing> billingOptional = billingRepository.findById(id);
-        return billingOptional.get();
+        if (billingOptional.isPresent()) {
+            return billingOptional.get();
+        } else {
+            throw new RecordNotFoundException();
+        }
     }
 
-
-
     @Override
-    public Billing saveBilling(Billing saveBilling) {
-        Billing billing = billingRepository.save(saveBilling);
-        return billing;
-    }
-
-
-
-    @Override
-    public Billing updateBilling(Long id, Billing billing) {
+    public Billing updateBilling(Long id, Billing newBilling) {
         Optional<Billing> billingOptional = billingRepository.findById(id);
-        billingOptional.get().setName(billing.getName());
-        billingOptional.get().setAmount(billing.getAmount());
-        billingOptional.get().setType(billing.getType());
-        return billingRepository.save(billingOptional.get());
+        if (billingOptional.isPresent()) {
+            Billing billing = billingOptional.get();
+            billing.setAmount(newBilling.getAmount());
+            billing.setType(newBilling.getType());
+            return billingRepository.save(billing);
+        } else {
+            throw new RecordNotFoundException();
+        }
     }
-
-
 
     @Override
     public void deleteBilling(Long id) {
         Optional<Billing> billingOptional = billingRepository.findById(id);
-        billingRepository.delete(billingOptional.get());
-
-
-
+        if (billingOptional.isPresent()) {
+            billingRepository.delete(billingOptional.get());
+        } else {
+            throw new RecordNotFoundException();
+        }
     }
+
 }
